requirements.txt
-i https://pypi.org/simple
faker==33.0.0; python_version >= '3.8'
mysql-connector==2.2.9
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
typing-extensions==4.12.2; python_version >= '3.8'
python-dateutil~=2.9.0.post0